requirements.txt
requests
pandas
matplotlib
python-telegram-bot>=20.0
gspread>=6.0
google-auth
google-auth-oauthlib
